Ref #1. Set the pid to usergroup system folder ID in Typo3.php, too.
TODO Where is the other place it is set? And which finally took effect?

Since this issue first arised a year ago, a lot has happened and user
groups now are set correctly in Aimeos. (see also commit 969358b)

In the case of the HAMAG webshop:
The users are put into `usergroup` with Id 1 (see Users folder in TYPO3
backend).
A configuration setting is set via `TypoScript` to assign all new frontend
users to this `group` by default.

While groups are set correctly now, all account, delivery related mails
do (still) not reach the customers because the mail address is empty.

This can be an issue of configuration or of code by @worlddevelopment,
Aimeos or TYPO3.